feat: refactor use-repeatable
