+++ lib/cjs/components/AddButton.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.AddButton = void 0;
 var react_1 = __importDefault(require("react"));
 function AddButton(_a) {
-    var onClick = _a.onClick, children = _a.children, addItem = _a.addItem, index = _a.index, newItem = _a.newItem, length = _a.length;
+    var onClick = _a.onClick, children = _a.children, addItem = _a.addItem, index = _a.index, newItem = _a.newItem, length = _a.length, className = _a.className;
     return (react_1.default.createElement("button", { onClick: function () {
             if (onClick) {
                 onClick();
@@ -17,6 +17,6 @@ function AddButton(_a) {
             else {
                 addItem();
             }
-        }, className: "repeatable-list__add-item-button" }, children));
+        }, className: "".concat(className, " repeatable-list__add-item-button") }, children));
 }
 exports.AddButton = AddButton;
