+++ lib/cjs/components/ClearButton.d.ts
@@ -1,0 +1,7 @@
+import React from "react";
+export declare function ClearButton({ onClick, children, removeAll, className, }: {
+    onClick?: () => void;
+    children: React.ReactNode;
+    removeAll: () => void;
+    className?: string;
+}): JSX.Element;
