+++ lib/cjs/components/MoveButton.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.MoveButton = void 0;
 var react_1 = __importDefault(require("react"));
 function MoveButton(_a) {
-    var onClick = _a.onClick, children = _a.children, moveItem = _a.moveItem, index = _a.index, direction = _a.direction, length = _a.length;
+    var onClick = _a.onClick, children = _a.children, moveItem = _a.moveItem, index = _a.index, direction = _a.direction, length = _a.length, className = _a.className;
     return (react_1.default.createElement("button", { onClick: function () {
             if (onClick) {
                 onClick();
@@ -23,6 +23,6 @@ function MoveButton(_a) {
             if (direction === "bottom") {
                 moveItem(index, length - 1);
             }
-        }, className: "repeatable-move-item-button" }, children));
+        }, className: "".concat(className, " repeatable-list__move-item-").concat(direction, "-button") }, children));
 }
 exports.MoveButton = MoveButton;
