+++ lib/cjs/modular-repeatable-list.js
@@ -23,13 +23,44 @@ var utilities_1 = require("@dnd-kit/utilities");
 var AddButton_1 = require("./components/AddButton");
 var RemoveButton_1 = require("./components/RemoveButton");
 var MoveButton_1 = require("./components/MoveButton");
+var ClearButton_1 = require("./components/ClearButton");
+function CardComp(_a) {
+    var sensors = _a.sensors, handleDragEnd = _a.handleDragEnd, items = _a.items, Card = _a.Card, removeItem = _a.removeItem, moveItem = _a.moveItem, addItem = _a.addItem, updateItem = _a.updateItem, removeAll = _a.removeAll;
+    return function () {
+        return (react_1.default.createElement(core_1.DndContext, { sensors: sensors, collisionDetection: core_1.closestCenter, onDragEnd: handleDragEnd },
+            react_1.default.createElement(sortable_1.SortableContext, { items: items, strategy: sortable_1.verticalListSortingStrategy }, items.map(function (item, n) { return (react_1.default.createElement(SortableCard, { key: item.id, id: item.id, Card: Card, cardProps: {
+                    // Props
+                    item: item,
+                    items: items,
+                    index: n,
+                    removeItem: removeItem,
+                    moveItem: moveItem,
+                    addItem: addItem,
+                    updateItem: updateItem,
+                    removeAll: removeAll,
+                    // Components
+                    AddButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem, className = _a.className;
+                        return (react_1.default.createElement(AddButton_1.AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length, className: className }));
+                    },
+                    RemoveButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, index = _a.index, className = _a.className;
+                        return (react_1.default.createElement(RemoveButton_1.RemoveButton, { onClick: onClick, children: children, removeItem: removeItem, index: index || n, className: className }));
+                    },
+                    MoveButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, direction = _a.direction, className = _a.className;
+                        return (react_1.default.createElement(MoveButton_1.MoveButton, { direction: direction, onClick: onClick, children: children, moveItem: moveItem, index: n, length: items.length, className: className }));
+                    },
+                } })); }))));
+    };
+}
 function RepeatableList(_a) {
     var initialState = _a.initialState, newItem = _a.newItem, onChange = _a.onChange, Card = _a.Card, Layout = _a.Layout;
     var _b = (0, _1.useRepeatable)({
         initialState: initialState || [],
         newItem: newItem,
         onChange: onChange,
-    }), items = _b.items, addItem = _b.addItem, removeItem = _b.removeItem, moveItem = _b.moveItem, updateItem = _b.updateItem;
+    }), items = _b.items, addItem = _b.addItem, removeItem = _b.removeItem, moveItem = _b.moveItem, updateItem = _b.updateItem, removeAll = _b.removeAll;
     function handleDragEnd(event) {
         var active = event.active, over = event.over;
         if (over === null) {
@@ -42,33 +73,23 @@ function RepeatableList(_a) {
         }
     }
     var sensors = (0, core_1.useSensors)((0, core_1.useSensor)(core_1.PointerSensor), (0, core_1.useSensor)(core_1.KeyboardSensor, { coordinateGetter: sortable_1.sortableKeyboardCoordinates }));
-    return (react_1.default.createElement(Layout, { items: items, addItem: addItem, removeItem: removeItem, moveItem: moveItem, updateItem: updateItem, AddButton: function (_a) {
-            var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem;
-            return (react_1.default.createElement(AddButton_1.AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length }));
-        }, Cards: function () { return (react_1.default.createElement(core_1.DndContext, { sensors: sensors, collisionDetection: core_1.closestCenter, onDragEnd: handleDragEnd },
-            react_1.default.createElement(sortable_1.SortableContext, { items: items, strategy: sortable_1.verticalListSortingStrategy }, items.map(function (item, n) { return (react_1.default.createElement(SortableCard, { key: item.id, id: item.id, Card: Card, cardProps: {
-                    // Props
-                    item: item,
-                    items: items,
-                    index: n,
-                    removeItem: removeItem,
-                    moveItem: moveItem,
-                    addItem: addItem,
-                    updateItem: updateItem,
-                    // Components
-                    AddButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem;
-                        return (react_1.default.createElement(AddButton_1.AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length }));
-                    },
-                    RemoveButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, index = _a.index;
-                        return (react_1.default.createElement(RemoveButton_1.RemoveButton, { onClick: onClick, children: children, removeItem: removeItem, index: index || n }));
-                    },
-                    MoveButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, direction = _a.direction;
-                        return (react_1.default.createElement(MoveButton_1.MoveButton, { direction: direction, onClick: onClick, children: children, moveItem: moveItem, index: n, length: items.length }));
-                    },
-                } })); })))); } }));
+    return (react_1.default.createElement(Layout, { items: items, addItem: addItem, removeItem: removeItem, moveItem: moveItem, updateItem: updateItem, removeAll: removeAll, AddButton: function (_a) {
+            var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem, className = _a.className;
+            return (react_1.default.createElement(AddButton_1.AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length, className: className }));
+        }, ClearButton: function (_a) {
+            var onClick = _a.onClick, children = _a.children, className = _a.className;
+            return (react_1.default.createElement(ClearButton_1.ClearButton, { onClick: onClick, children: children, className: className, removeAll: removeAll }));
+        }, Cards: CardComp({
+            sensors: sensors,
+            handleDragEnd: handleDragEnd,
+            items: items,
+            Card: Card,
+            removeItem: removeItem,
+            moveItem: moveItem,
+            addItem: addItem,
+            updateItem: updateItem,
+            removeAll: removeAll,
+        }) }));
 }
 exports.RepeatableList = RepeatableList;
 function SortableCard(_a) {
