+++ lib/esm/components/AddButton.js
@@ -1,6 +1,6 @@
 import React from "react";
 export function AddButton(_a) {
-    var onClick = _a.onClick, children = _a.children, addItem = _a.addItem, index = _a.index, newItem = _a.newItem, length = _a.length;
+    var onClick = _a.onClick, children = _a.children, addItem = _a.addItem, index = _a.index, newItem = _a.newItem, length = _a.length, className = _a.className;
     return (React.createElement("button", { onClick: function () {
             if (onClick) {
                 onClick();
@@ -11,5 +11,5 @@ export function AddButton(_a) {
             else {
                 addItem();
             }
-        }, className: "repeatable-list__add-item-button" }, children));
+        }, className: "".concat(className, " repeatable-list__add-item-button") }, children));
 }
