+++ lib/esm/components/MoveButton.js
@@ -1,6 +1,6 @@
 import React from "react";
 export function MoveButton(_a) {
-    var onClick = _a.onClick, children = _a.children, moveItem = _a.moveItem, index = _a.index, direction = _a.direction, length = _a.length;
+    var onClick = _a.onClick, children = _a.children, moveItem = _a.moveItem, index = _a.index, direction = _a.direction, length = _a.length, className = _a.className;
     return (React.createElement("button", { onClick: function () {
             if (onClick) {
                 onClick();
@@ -17,5 +17,5 @@ export function MoveButton(_a) {
             if (direction === "bottom") {
                 moveItem(index, length - 1);
             }
-        }, className: "repeatable-move-item-button" }, children));
+        }, className: "".concat(className, " repeatable-list__move-item-").concat(direction, "-button") }, children));
 }
