+++ lib/esm/components/RemoveButton.js
@@ -1,10 +1,10 @@
 import React from "react";
 export function RemoveButton(_a) {
-    var onClick = _a.onClick, children = _a.children, removeItem = _a.removeItem, index = _a.index;
+    var onClick = _a.onClick, children = _a.children, removeItem = _a.removeItem, index = _a.index, className = _a.className;
     return (React.createElement("button", { onClick: function () {
             if (onClick) {
                 onClick();
             }
             removeItem(index);
-        }, className: "repeatable-remove-item-button" }, children));
+        }, className: "".concat(className, " repeatable-list__remove-item-button") }, children));
 }
