+++ lib/esm/modular-repeatable-list.js
@@ -17,13 +17,44 @@ import { CSS } from "@dnd-kit/utilities";
 import { AddButton } from "./components/AddButton";
 import { RemoveButton } from "./components/RemoveButton";
 import { MoveButton } from "./components/MoveButton";
+import { ClearButton } from "./components/ClearButton";
+function CardComp(_a) {
+    var sensors = _a.sensors, handleDragEnd = _a.handleDragEnd, items = _a.items, Card = _a.Card, removeItem = _a.removeItem, moveItem = _a.moveItem, addItem = _a.addItem, updateItem = _a.updateItem, removeAll = _a.removeAll;
+    return function () {
+        return (React.createElement(DndContext, { sensors: sensors, collisionDetection: closestCenter, onDragEnd: handleDragEnd },
+            React.createElement(SortableContext, { items: items, strategy: verticalListSortingStrategy }, items.map(function (item, n) { return (React.createElement(SortableCard, { key: item.id, id: item.id, Card: Card, cardProps: {
+                    // Props
+                    item: item,
+                    items: items,
+                    index: n,
+                    removeItem: removeItem,
+                    moveItem: moveItem,
+                    addItem: addItem,
+                    updateItem: updateItem,
+                    removeAll: removeAll,
+                    // Components
+                    AddButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem, className = _a.className;
+                        return (React.createElement(AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length, className: className }));
+                    },
+                    RemoveButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, index = _a.index, className = _a.className;
+                        return (React.createElement(RemoveButton, { onClick: onClick, children: children, removeItem: removeItem, index: index || n, className: className }));
+                    },
+                    MoveButton: function (_a) {
+                        var onClick = _a.onClick, children = _a.children, direction = _a.direction, className = _a.className;
+                        return (React.createElement(MoveButton, { direction: direction, onClick: onClick, children: children, moveItem: moveItem, index: n, length: items.length, className: className }));
+                    },
+                } })); }))));
+    };
+}
 function RepeatableList(_a) {
     var initialState = _a.initialState, newItem = _a.newItem, onChange = _a.onChange, Card = _a.Card, Layout = _a.Layout;
     var _b = useRepeatable({
         initialState: initialState || [],
         newItem: newItem,
         onChange: onChange,
-    }), items = _b.items, addItem = _b.addItem, removeItem = _b.removeItem, moveItem = _b.moveItem, updateItem = _b.updateItem;
+    }), items = _b.items, addItem = _b.addItem, removeItem = _b.removeItem, moveItem = _b.moveItem, updateItem = _b.updateItem, removeAll = _b.removeAll;
     function handleDragEnd(event) {
         var active = event.active, over = event.over;
         if (over === null) {
@@ -36,33 +67,23 @@ function RepeatableList(_a) {
         }
     }
     var sensors = useSensors(useSensor(PointerSensor), useSensor(KeyboardSensor, { coordinateGetter: sortableKeyboardCoordinates }));
-    return (React.createElement(Layout, { items: items, addItem: addItem, removeItem: removeItem, moveItem: moveItem, updateItem: updateItem, AddButton: function (_a) {
-            var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem;
-            return (React.createElement(AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length }));
-        }, Cards: function () { return (React.createElement(DndContext, { sensors: sensors, collisionDetection: closestCenter, onDragEnd: handleDragEnd },
-            React.createElement(SortableContext, { items: items, strategy: verticalListSortingStrategy }, items.map(function (item, n) { return (React.createElement(SortableCard, { key: item.id, id: item.id, Card: Card, cardProps: {
-                    // Props
-                    item: item,
-                    items: items,
-                    index: n,
-                    removeItem: removeItem,
-                    moveItem: moveItem,
-                    addItem: addItem,
-                    updateItem: updateItem,
-                    // Components
-                    AddButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem;
-                        return (React.createElement(AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length }));
-                    },
-                    RemoveButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, index = _a.index;
-                        return (React.createElement(RemoveButton, { onClick: onClick, children: children, removeItem: removeItem, index: index || n }));
-                    },
-                    MoveButton: function (_a) {
-                        var onClick = _a.onClick, children = _a.children, direction = _a.direction;
-                        return (React.createElement(MoveButton, { direction: direction, onClick: onClick, children: children, moveItem: moveItem, index: n, length: items.length }));
-                    },
-                } })); })))); } }));
+    return (React.createElement(Layout, { items: items, addItem: addItem, removeItem: removeItem, moveItem: moveItem, updateItem: updateItem, removeAll: removeAll, AddButton: function (_a) {
+            var onClick = _a.onClick, children = _a.children, index = _a.index, newItem = _a.newItem, className = _a.className;
+            return (React.createElement(AddButton, { onClick: onClick, children: children, addItem: addItem, index: index, newItem: newItem, length: items.length, className: className }));
+        }, ClearButton: function (_a) {
+            var onClick = _a.onClick, children = _a.children, className = _a.className;
+            return (React.createElement(ClearButton, { onClick: onClick, children: children, className: className, removeAll: removeAll }));
+        }, Cards: CardComp({
+            sensors: sensors,
+            handleDragEnd: handleDragEnd,
+            items: items,
+            Card: Card,
+            removeItem: removeItem,
+            moveItem: moveItem,
+            addItem: addItem,
+            updateItem: updateItem,
+            removeAll: removeAll,
+        }) }));
 }
 function SortableCard(_a) {
     var id = _a.id, Card = _a.Card, cardProps = _a.cardProps;
