+++ src/repeatable-list.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from "react";
+import React from "react";
 import { useRepeatable } from ".";
 import {
   DndContext,
@@ -9,101 +9,32 @@ import {
   closestCenter,
   DragEndEvent,
 } from "@dnd-kit/core";
-import { CSS } from "@dnd-kit/utilities";
 import {
   SortableContext,
   sortableKeyboardCoordinates,
-  verticalListSortingStrategy,
   useSortable,
+  verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
+import { CSS } from "@dnd-kit/utilities";
+import { SyntheticListenerMap } from "@dnd-kit/core/dist/hooks/utilities";
+import { AddButton } from "./components/AddButton";
+import { RemoveButton } from "./components/RemoveButton";
+import { MoveButton } from "./components/MoveButton";
+import { ClearButton } from "./components/ClearButton";
 
-import { AddItemButton } from "./components/AddItemButton";
-import { RepeatableListProps, SortableCardProps } from "./types";
-
-// const defaultStyles = `
-//   .repeatable-list__control-button-container {
-//     display: flex;
-//     flex-grow: 0;
-//     gap: 3px;
-//   }
-//   .repeatable-list__remove-item-button {
-//     width: 100px;
-//     height: 30px;
-//     padding: 5px;
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     margin: 0;
-//   }
-//   .repeatable-list__reorder-item-button {
-//     width: 30px;
-//     height: 30px;
-//     padding: 5px;
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     margin: 0;
-//   }
-
-//   .repeatable-list__add-item-button {
-//     height: 30px;
-//     padding: 5px;
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     margin: 10px auto;
-//   }
-
-//   .repeatable-list__card {
-//     display: flex;
-//     gap: 5px;
-//   }
-
-//   .repeatable-list__list-item {
-//     display: flex;
-//     flex-grow: 1;
-//   }
-
-//   .repeatable-list__drag-handle {
-//     width: 30px;
-//     height: 30px;
-//     padding: 5px;
-//     &:hover {
-//       cursor: grab;
-//     }
-//   }
-
-//   .repeatable-list__drag-handle-icon {
-//     width: 100%;
-//     height: 100%;
-//   }
-// `;
-
-const RepeatableList = <T extends {}>(
-  props: RepeatableListProps<T> & {
-    children?: ({ test }: { test: string }) => ReactNode;
-  }
-) => {
-  const {
-    initialState,
-    listItem,
-    onChange,
-    newItem,
-    AddItemComponent,
-    RemoveItemComponent,
-    ReorderItemDownComponent,
-    ReorderItemUpComponent,
-    DragHandleComponent,
-    composeInnerComponents,
-    composeOuterComponents,
-    draggable,
-  } = props;
-
-  const { items, addItem, removeItem, moveItem, updateItem } = useRepeatable({
-    initialState: initialState || [],
-    newItem: newItem,
-    onChange: onChange,
-  });
+function RepeatableList<T>({
+  initialState,
+  newItem,
+  onChange,
+  Card,
+  Layout,
+}: RepeatableListProps<T>) {
+  const { items, addItem, removeItem, moveItem, updateItem, removeAll } =
+    useRepeatable<T>({
+      initialState: initialState || [],
+      newItem: newItem,
+      onChange: onChange,
+    });
 
   function handleDragEnd(event: DragEndEvent) {
     const { active, over } = event;
@@ -122,16 +53,36 @@ const RepeatableList = <T extends {}>(
     useSensor(KeyboardSensor, { coordinateGetter: sortableKeyboardCoordinates })
   );
 
-  if (composeOuterComponents) {
-    return (
-      <div className="repeatable-list__wrapper">
-        {composeOuterComponents({
-          repeatable: { items, addItem, removeItem, moveItem, updateItem },
-          AddItemButton: AddItemButton({
-            CustomComponent: AddItemComponent,
-            addItem: addItem,
-          }),
-          InnerComponents: (
+  return (
+    <div className="repeatable-list__wrapper">
+      <Layout
+        items={items}
+        addItem={addItem}
+        removeItem={removeItem}
+        moveItem={moveItem}
+        updateItem={updateItem}
+        removeAll={removeAll}
+        AddButton={({ onClick, children, index, newItem, className }) => (
+          <AddButton
+            onClick={onClick}
+            children={children}
+            addItem={addItem}
+            index={index}
+            newItem={newItem}
+            length={items.length}
+            className={className}
+          />
+        )}
+        ClearButton={({ onClick, children, className }) => (
+          <ClearButton
+            onClick={onClick}
+            children={children}
+            className={className}
+            removeAll={removeAll}
+          />
+        )}
+        Cards={() => (
+          <div className="repeatable-list__cards">
             <DndContext
               sensors={sensors}
               collisionDetection={closestCenter}
@@ -141,252 +92,182 @@ const RepeatableList = <T extends {}>(
                 items={items}
                 strategy={verticalListSortingStrategy}
               >
-                {items.map((item, n) => {
-                  return (
-                    <SortableCard
-                      composeInnerComponents={composeInnerComponents}
-                      AddItemComponent={AddItemComponent}
-                      RemoveItemComponent={RemoveItemComponent}
-                      ReorderItemUpComponent={ReorderItemUpComponent}
-                      ReorderItemDownComponent={ReorderItemDownComponent}
-                      DragHandleComponent={DragHandleComponent}
-                      item={item}
-                      items={items}
-                      key={item.id}
-                      n={n}
-                      addItem={addItem}
-                      removeItem={removeItem}
-                      updateItem={(i) => {
-                        updateItem(n, i);
-                      }}
-                      moveItem={moveItem}
-                      listItem={listItem}
-                      showReorderButtons={props.showReorderButtons}
-                      draggable={draggable}
-                      repeatable={{
-                        items,
-                        addItem,
-                        removeItem,
-                        moveItem,
-                        updateItem,
-                      }}
-                    />
-                  );
-                })}
+                {items.map((item, n) => (
+                  <SortableCard
+                    key={item.id}
+                    id={item.id}
+                    Card={Card}
+                    cardProps={{
+                      // Props
+                      item,
+                      items,
+                      index: n,
+                      removeItem,
+                      moveItem,
+                      addItem,
+                      updateItem,
+                      removeAll,
+
+                      // Components
+                      AddButton: ({
+                        onClick,
+                        children,
+                        index,
+                        newItem,
+                        className,
+                      }) => (
+                        <AddButton
+                          onClick={onClick}
+                          children={children}
+                          addItem={addItem}
+                          index={index}
+                          newItem={newItem}
+                          length={items.length}
+                          className={className}
+                        />
+                      ),
+                      RemoveButton: ({
+                        onClick,
+                        children,
+                        index,
+                        className,
+                      }) => (
+                        <RemoveButton
+                          onClick={onClick}
+                          children={children}
+                          removeItem={removeItem}
+                          index={index || n}
+                          className={className}
+                        />
+                      ),
+                      MoveButton: ({
+                        onClick,
+                        children,
+                        direction,
+                        className,
+                      }) => (
+                        <MoveButton
+                          direction={direction}
+                          onClick={onClick}
+                          children={children}
+                          moveItem={moveItem}
+                          index={n}
+                          length={items.length}
+                          className={className}
+                        />
+                      ),
+                    }}
+                  />
+                ))}
               </SortableContext>
             </DndContext>
-          ),
-        })}
-      </div>
-    );
-  } else {
-    return (
-      <div className="repeatable-list__wrapper">
-        <DndContext
-          sensors={sensors}
-          collisionDetection={closestCenter}
-          onDragEnd={handleDragEnd}
-        >
-          <SortableContext items={items} strategy={verticalListSortingStrategy}>
-            {items.map((item, n) => {
-              return (
-                <SortableCard
-                  composeInnerComponents={composeInnerComponents}
-                  AddItemComponent={AddItemComponent}
-                  RemoveItemComponent={RemoveItemComponent}
-                  ReorderItemUpComponent={ReorderItemUpComponent}
-                  ReorderItemDownComponent={ReorderItemDownComponent}
-                  DragHandleComponent={DragHandleComponent}
-                  item={item}
-                  items={items}
-                  key={item.id}
-                  n={n}
-                  addItem={addItem}
-                  removeItem={removeItem}
-                  updateItem={(i) => {
-                    updateItem(n, i);
-                  }}
-                  moveItem={moveItem}
-                  listItem={listItem}
-                  showReorderButtons={props.showReorderButtons}
-                  draggable={draggable}
-                  repeatable={{
-                    items,
-                    addItem,
-                    removeItem,
-                    moveItem,
-                    updateItem,
-                  }}
-                />
-              );
-            })}
-          </SortableContext>
-        </DndContext>
+          </div>
+        )}
+      />
+    </div>
+  );
+}
 
-        <AddItemButton CustomComponent={AddItemComponent} addItem={addItem} />
-      </div>
-    );
-  }
-};
-
-const SortableCard = <T extends Object>({
-  item,
-  items,
-  n,
-  addItem,
-  removeItem,
-  moveItem,
-  listItem,
-  updateItem,
-  repeatable,
-
-  //components
-  /**Use a custom component for the "Remove Item" button.
-   * Additional functions can be added alongside the default function.*/
-  RemoveItemComponent,
-  /**Use a custom component for the reorder up button.
-   * Additional functions can be added alongside the default function.*/
-  ReorderItemUpComponent,
-  /**Use a custom component for the reorder down button.
-   * Additional functions can be added alongside the default function.*/
-  ReorderItemDownComponent,
-  /**Use a custom component for the drag handle.
-   * Additional functions can be added alongside the default function.*/
-  DragHandleComponent,
-  AddItemComponent,
-  composeInnerComponents,
-
-  // options
-  /**Enable up and down arrow buttons to reorder items*/
-  showReorderButtons = true,
-  /**Enable drag handles to reorder items*/
-  draggable = true,
-}: SortableCardProps<T>) => {
+function SortableCard<T>({ id, Card, cardProps }: SortableCardProps<T>) {
   const { attributes, listeners, setNodeRef, transform, transition } =
-    useSortable({ id: item.id });
+    useSortable({ id: id });
 
-  const DragHandle = () => {
-    if (DragHandleComponent) {
-      return (
-        <b className="repeatable-list__drag-handle" {...listeners}>
-          <DragHandleComponent />
-        </b>
-      );
-    } else {
-      return (
-        <b className="repeatable-list__drag-handle" {...listeners}>
-          ≡
-        </b>
-      );
-    }
-  };
-
-  const RemoveItemButton = () => {
-    if (RemoveItemComponent) {
-      return (
-        <RemoveItemComponent
-          onClick={(e) => {
-            removeItem(n);
-            e.preventDefault();
-          }}
-        />
-      );
-    } else {
-      return (
-        <button
-          className="repeatable-list__remove-item-button"
-          onClick={(e) => {
-            removeItem(n);
-            e.preventDefault();
-          }}
-        >
-          x
-        </button>
-      );
-    }
-  };
-
-  const ReorderItemButton = (direction: "up" | "down") => {
-    if (ReorderItemDownComponent) {
-      return (
-        <ReorderItemDownComponent
-          onClick={(e) => {
-            moveItem(n, n + 1);
-            e.preventDefault();
-          }}
-        />
-      );
-    } else if (ReorderItemUpComponent) {
-      return (
-        <ReorderItemUpComponent
-          onClick={(e) => {
-            moveItem(n, n - 1);
-            e.preventDefault();
-          }}
-        />
-      );
-    } else {
-      return (
-        <button
-          className={`repeatable-reorder-item-${direction}-button`}
-          onClick={(e) => {
-            moveItem(n, direction === "up" ? n - 1 : n + 1);
-            e.preventDefault();
-          }}
-        >
-          {direction === "up" ? "^" : "v"}
-        </button>
-      );
-    }
-  };
-
-  if (composeInnerComponents) {
-    return (
-      <div
-        ref={setNodeRef}
-        {...attributes}
-        style={{ transform: CSS.Transform.toString(transform), transition }}
-      >
-        {composeInnerComponents({
-          repeatable: { ...repeatable, index: n, item, items },
-          DragHandle: DragHandle(),
-          ListItem: listItem(item, updateItem, n, items),
-          RemoveItemButton: RemoveItemButton(),
-          ReorderItemDownButton: ReorderItemButton("down"),
-          ReorderItemUpButton: ReorderItemButton("up"),
-          AddItemButton: AddItemButton({
-            CustomComponent: AddItemComponent,
-            addItem: addItem,
-          }),
-        })}
-      </div>
-    );
-  } else {
-    return (
-      <div
-        className={`repeatable-list__card`}
-        ref={setNodeRef}
-        style={{ transform: CSS.Transform.toString(transform), transition }}
-        {...attributes}
-      >
-        {draggable && <DragHandle />}
-
-        <div className="repeatable-list__list-item">
-          {listItem(item, updateItem, n, items)}
-        </div>
-        <div className="repeatable-list__control-button-container">
-          <RemoveItemButton />
-
-          {showReorderButtons && (
-            <>
-              {ReorderItemButton("up")}
-              {ReorderItemButton("down")}
-            </>
-          )}
-        </div>
-      </div>
-    );
-  }
-};
+  return (
+    <div
+      className="repeatable-list__card"
+      ref={setNodeRef}
+      {...attributes}
+      style={{ transform: CSS.Transform.toString(transform), transition }}
+    >
+      <Card
+        {...cardProps}
+        DragHandle={({ children }: { children: React.ReactNode }) => (
+          <div
+            style={{ cursor: "grab" }}
+            className="repeatable-list__drag-handle"
+            {...listeners}
+          >
+            {children}
+          </div>
+        )}
+        dragHandleListeners={listeners}
+      />
+    </div>
+  );
+}
 
 export { RepeatableList };
+
+interface RepeatableListProps<T> {
+  initialState: T[];
+  newItem: T;
+  onChange?: (items: (T & { id: string })[]) => void;
+  Card: React.FC<CardProps<T>>;
+  Layout: React.FC<LayoutProps<T>>;
+}
+
+interface SortableCardProps<T> {
+  id: string;
+  Card: React.FC<CardProps<T>>;
+  cardProps: Omit<CardProps<T>, "dragHandleListeners">;
+}
+
+interface CardProps<T> {
+  // Props
+  item: T & { id: string };
+  items: (T & { id: string })[];
+  index: number;
+  removeItem: (n: number) => void;
+  moveItem: (from: number, to: number) => void;
+  addItem: (item?: T, n?: number) => void;
+  updateItem: (n: number, item: T) => void;
+  removeAll: () => void;
+  dragHandleListeners?: SyntheticListenerMap;
+
+  // Pre-made components
+  DragHandle?: React.FC<{ children?: React.ReactNode }>;
+  AddButton?: React.FC<{
+    children?: React.ReactNode;
+    onClick?: () => void;
+    index?: number;
+    newItem?: T;
+    className?: string;
+  }>;
+  RemoveButton?: React.FC<{
+    children?: React.ReactNode;
+    onClick?: () => void;
+    index?: number;
+    className?: string;
+  }>;
+  MoveButton?: React.FC<{
+    children?: React.ReactNode;
+    onClick?: () => void;
+    direction: "up" | "down" | "top" | "bottom";
+    className?: string;
+  }>;
+}
+
+interface LayoutProps<T> {
+  // Props
+  items: (T & { id: string })[];
+  removeItem: () => void;
+  moveItem: (from: number, to: number) => void;
+  addItem: (item?: T, n?: number) => void;
+  updateItem: (n: number, item: T) => void;
+  removeAll: () => void;
+
+  // Components
+  Cards: React.FC<{}>;
+  AddButton?: React.FC<{
+    children?: React.ReactNode;
+    onClick?: () => void;
+    index?: number;
+    newItem?: T;
+    className?: string;
+  }>;
+  ClearButton?: React.FC<{
+    children?: React.ReactNode;
+    onClick?: () => void;
+    className?: string;
+  }>;
+}
